refactor: unify SbtServer and ScalaCliServer
